оставить удаление группы только владельцу

+++ src/widgets/Chats/ui/Chats.js
@@ -1,5 +1,5 @@
 import { eventBus } from "../../../shared/modules/EventBus/EventBus.js";
-import { User } from "../../../entities/User/model/User.js";
+import { currentUser, User } from "../../../entities/User/model/User.js";
 import { api } from "../../../shared/api/api.js";
 import { deleteChat } from "../../../entities/Chat/api/api.js";
 import { createChat } from "../../../entities/Chat/api/api.js";
@@ -31,12 +31,36 @@ class Chats {
 
         if (this.chats !== null && this.chats !== undefined) {
             for (const chat of this.chats) {
-                chats.push({
+                console.log("chat:", chat);
+
+                let isOwner = true;
+                if (chat["type"] == "group") {
+                    const responseBody = await api.get(`/chat/${chat["id"]}`);
+                    console.log("group response:", responseBody);
+
+                    const users = responseBody["data"]["users"];
+                    console.log("group users:", users);
+
+                    const currentMember = users.find(
+                        (item) =>
+                            item["username"] === currentUser.getUsername(),
+                    );
+                    console.log("current member:", currentMember);
+
+                    if (currentMember["role"] === "member") {
+                        isOwner = false;
+                    }
+                }
+
+                const chatInfo = {
                     title: chat.title,
                     // lastMessage:
                     avatarSrc: await getAvatar(chat.avatar_path),
                     chatId: chat.id,
-                });
+                    isOwner: isOwner,
+                };
+
+                chats.push(chatInfo);
             }
         }
 
@@ -84,13 +108,15 @@ class Chats {
 
             // Удалить чат
             const deleteChatButton = chatElement.querySelector(".button");
-            deleteChatButton.addEventListener("click", async (event) => {
-                event.preventDefault();
-                event.stopPropagation();
+            if (deleteChatButton !== null) {
+                deleteChatButton.addEventListener("click", async (event) => {
+                    event.preventDefault();
+                    event.stopPropagation();
 
-                await deleteChat(chatModel.id);
-                eventBus.emit("chat is deleted");
-            });
+                    await deleteChat(chatModel.id);
+                    eventBus.emit("chat is deleted");
+                });
+            }
         }
 
         // Обработчик клика по кнопке меню
@@ -169,29 +195,7 @@ class Chats {
         const newDialog = newChatPopoverElement.querySelector("#new-dialog");
         newDialog.addEventListener("click", async (event) => {
             event.preventDefault();
-            // const username = prompt(
-            //     "Введите username пользователя, которому Вы хотите написать",
-            // );
-
-            // if (username !== null) {
-            //     const responseBody = await api.get(`/profile/${username}`);
-
-            //     if (responseBody.status === true) {
-            //         const chatData = {
-            //             type: "dialog",
-            //             dialog_user: username,
-            //             title: "1",
-            //         };
-            //         await createChat(chatData);
-
-            //         chatWebSocket.reconnect();
-
-            //         eventBus.emit("new chat is created");
-            //     } else {
-            //         alert(`Пользователь с username "${username}" не найден`);
-            //     }
-            // }
-            eventBus.emit("chats -> new dialog")
+            eventBus.emit("chats -> new dialog");
         });
 
         // Новая группа
